feat(frontend): Update refactor
Added a wrapper above update form to make sure it only
ever displays once the data has been fetched

IssueID #
