+++ src/functionality/blogging/blogpost.component.tsx
@@ -6,6 +6,8 @@ import usePost from './usePost';
 
 const BlogPost: React.FC = (): ReactElement => {
   const { id } = useParams();
+  console.log('blogpost id', id);
+  console.log('blogpost typeof', typeof id);
   const navigate = useNavigate();
 
   const { isLoading, isError, data: post, error } = usePost(id);
